requirements.txt
Django==1.6.5
MySQL-python==1.2.5
argparse==1.2.1
dicttoxml==1.5.5
django-mptt==0.6.1
wsgiref==0.1.2
